Give the interactive chat entrypoint the same provisioning surface as batch
Host bind mounts and the model-turn budget were reachable only through the
batch CLI. The interactive stdio host is a second entrypoint over the same
engine, and it parsed three flags and threw on everything else — so an embedder
could provision a batch run and not a chat session, with nothing in the types
saying why.

Loom shipped cf-harness interactive chat on that gap. Chat ran with no /loom
mount and no Page RPC queue, so page authoring silently disappeared from chat
while it kept working for wishes. The 8-turn default was the same gap: batch
exposes --max-model-turns, interactive exposed nothing, so an ordinary
four-file read failed 2 of 3 runs with "prompt loop exceeded max model turns".

The mount grammar, parser and engine mapping move out of cli.ts into
host-mounts.ts, and both entrypoints use them. Keeping the parser inside the
batch entrypoint is how the second entrypoint ended up unable to mount
anything; the module says so, so the next entrypoint calls parseHostMountSpecs
rather than inventing a second way in.

No contract or protocol change was needed: CreateHarnessPromptLoopOptions
already extends CreateHarnessEngineOptions, and the loop builds its engine from
those options, so additionalMounts and maxModelTurns reach the sandbox through
basePromptLoopOptions, which the interactive service already spreads per turn.

Tests: new coverage for the added flags, the zero-turn rejection, and that
unknown args are still refused. 742 cf-harness tests pass, including the 147
batch CLI tests that exercise the lifted parser.
